Make tensor of indices a LongTensor
Indices get multiplied with strides in operations like aten::index_put. A
32-bit index can lead to a silent overflow, since the Torch Inductor does not
do any overflow checking on its own.

This commit fixes GitHub issue #2430

+++ torchtitan/models/common/moe/kernels.py
@@ -80,7 +80,7 @@ def fill_indices_wrapper(
 ):
     # preallocate output
     permuted_indices = torch.full(
-        (max_len,), -1, dtype=torch.int32, device=tokens_per_expert_group.device
+        (max_len,), -1, dtype=torch.int64, device=tokens_per_expert_group.device
     )
 
     # write offsets is per local expert...
@@ -118,7 +118,7 @@ def fill_indices_cpu(
     permuted_indices = torch.full(
         (max_len,),
         -1,
-        dtype=torch.int32,
+        dtype=torch.int64,
     )  # device=device)
     # Fill the permuted indices
     # For each local expert
